Crate origen

Re-exports

pub extern crate origen_metal;

pub use self::core::status::Operation;

pub use self::generator::utility::transaction::Action as TransactionAction;

pub use self::generator::utility::transaction::Transaction;

pub use self::core::dut::Dut;

pub use self::core::producer::Producer;

pub use self::core::status::app_present;

pub use self::core::status::in_app_invocation;

pub use self::core::status::in_global_invocation;

pub use self::core::tester::Capture;

pub use self::core::tester::Overlay;

pub use self::core::tester::Tester;

pub use self::services::Services;

pub use self::utility::sessions::setup_sessions;

pub use self::utility::sessions::with_app_session;

pub use self::utility::sessions::with_app_session_group;

pub use self::core::frontend::callbacks as CALLBACKS;

pub use self::core::frontend::emit_callback;

pub use self::core::frontend::with_frontend;

pub use self::core::frontend::with_frontend_app;

pub use self::core::frontend::with_optional_frontend;

pub use origen_metal as om;

Modules

built_info

core

generator

precludes

services

standards

testers

utility

Macros

add_children

add_reg

add_reg_32bit

backend_lookup_error

Returns a error stating that the backend doesn’t have an expected ID. This signals a bug somewhere and should only be used when we’re expecting something to exists.

clean_target

duplicate_error

Returns an Origen Error with pre-formatted message complaining that something already exists.

exit_on_bad_config

field

get_bc_for

get_reg

get_reg_as_bc

lookup_error

push_pin_actions

some_hard_reset_val

swd_ok

text

text_line

Structs

DUT

The current device model, containing all metadata about hierarchy, regs, pins, specs, timing, etc. and responsible for maintaining the current state of the DUT (regs, pins, etc.)

Error

FRONTEND

FeatureReturn

LOGGER

ORIGEN_CONFIG

Provides configuration information derived from origen.toml files found in the Origen installation and application file system paths

ORIGEN_CONFIG_METADATA

Outcome

PRODUCER

Producer

SERVICES

Services owned by the current DUT, stored as a separate collection to avoid having to get a mutable ref on the DUT if the service needs mutation

STATUS

Provides status information derived from the runtime environment, e.g. if an app is present If an app is present then its Application struct is stored in here. Things like the current output and reference directories should be derived from here.

TEST

Storage for the current test (pattern)

TESTER

The global tester model.

TypedValueMap

Wrapper around an indexmap of typed values.

Enums

DataStoreFeature

DsFeature

TypedValue

Value

Constants

MODES

The available Origen runtime modes

Statics

FE_CAT_NAME__LDAPS

Traits

DataStoreCategoryFrontendAPI

Allows for arbitrary “data sources”. This could be, for example, a database (any), LDAP, just a random file, etc. Just need a few methods implemented for frontend <-> backend integration. Others are available but are optional.

DataStoreFrontendAPI

DsAPI

DsCatAPI

Allows for arbitrary “data sources”. This could be, for example, a database (any), LDAP, just a random file, etc. Just need a few methods implemented for frontend <-> backend integration. Others are available but are optional.

RcAPI

RevisionControlFrontendAPI

Functions

app

clean_mode

Sanitizes the given mode string and returns it, but will exit the process if it is invalid

dut

initialize

This is called immediately upon Origen booting

origen_config_metadata

prepare_for_target_load

This will be called immediately before loading a fresh set of targets. Everything required to clear previous state from the existing targets should be initiated from here.

producer

services

start_new_test

Clears the current test (pattern) AST and starts a new one, this will be called by the producer before loading the next pattern source file

tester

trace_error

with_current_job

Execute the given function with a reference to the current job. Returns an error if there is no current job, otherwise the result of the given function.

with_current_job_mut

Execute the given function with a mutable reference to the current job. Returns an error if there is no current job, otherwise the result of the given function.

Type Aliases

Result