origen::commands

Module _prelude

Re-exports

pub use super::launch_as;

pub use crate::framework::core_cmds::SubCmd;

pub use crate::framework::CmdHelps;

pub use crate::framework::CmdSrc;

pub use crate::framework::Extensions;

pub use crate::framework::Plugins;

pub use crate::origen_fe_available;

pub use crate::output_dir_opt;

pub use crate::ref_dir_opt;

pub use crate::req_sv_arg;

pub use crate::sv_opt;

pub use crate::core_subcmd;

pub use crate::core_subcmd__no_exts__no_app_opts;

pub use crate::gen_core_cmd_funcs;

pub use crate::gen_core_cmd_funcs__no_exts__no_app_opts;

pub use crate::gen_simple_run_func;

pub use crate::print_subcmds_available_msg;

pub use crate::unreachable_invalid_subc;

pub use clap_arg_actions::*;

Modules

clap_arg_actions

Structs

Arg

The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

Command

Build a command-line interface.

IndexMap

A hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

Type Aliases

App

Result

RunInput