Module origen::commands::proj

Modules

bom

group

package

Statics

BOM_FILE

README_FILE

Functions

error_and_exit

get_dir_or_pwd
If validate is not true then the path returned may not be absolute and the caller is responsible for handling that (if required) after then have created it

get_package_ids_from_args
Resolves package and group IDs in the args to a vector of package IDs. If a given ID does not match a known package or group the process will be exited with an error. Optionally return all packages if no packages arg given.

pwd

run

validate_path
Will exit and print an error message if the given path reference is invalid. Caller must specify whether they want the path to exist or not and whether they expect it to be a file or a dir (if applicable).