Struct origen::commands::proj::bom::BOM

pub struct BOM {
    pub meta: Meta,
    pub files: Vec<PathBuf>,
    pub packages: IndexMap<String, Package>,
    pub links: IndexMap<String, String>,
    pub groups: IndexMap<String, Group>,
}

Fields
meta: Metafiles: Vec<PathBuf>packages: IndexMap<String, Package>links: IndexMap<String, String>groups: IndexMap<String, Group>
Implementations

impl BOM

pub fn for_dir(dir: &Path) -> BOM

pub fn resolve_ids(&self, ids: Vec<&str>) -> Result<Vec<String>>

Returns the package IDs matching the given package or group IDs. An error will be returned if one of the given IDs does not match a known package or group.

fn merge(&mut self, bom: BOM, source: &Path)

fn apply_group_versions(&mut self)

Applies the version from a group to any of its child packages that doesn't have one

fn copy_group_packages(&mut self, bom: &BOM)

Child BOMs are not allowed to redefine what packages are in a group, this is used to copy the existing definitions from a parent BOM. It is done separately from merging to support the version resolution flow.

fn validate(&self)

pub fn is_workspace(&self) -> bool

Returns true if the BOM belongs to a workspace

pub fn root(&self) -> &Path

Returns an absolute path to the workspace top-level directory

pub fn create_links(&self, force: bool) -> Result<bool>