origen

Module core

Modules

application

config

Collects the main Origen configuration options from all origen.toml files found in the application and Origen CLI installation file system paths

dut

file_handler

The file_handler is responsible for processing the file arguments supplied to Origen commands from the CLI. It provides methods for consumers to retreive one file at a time or all files at once and seamlessly opens up lists to get to the inidividual files inside.

frontend

lsf

This module is responsible for abstracting and managing job submissions to LSF

model

producer

status

term

tester