origen::core::application

Module target

A default target/env for an application can be set by config/application.toml, this can then be overridden by the user by setting a temporary workspace default via the “origen t” and “origen e” commands. These commands store the user’s selections in .origen/application.toml and this file should NOT be checked into revision control.

The target can be further overridden for a particular origen command invocation via the -t and -e options, or programmatically within the application code, however that is all handled on the front end in Python code.

Functions

add

Enables additional targets in the workspace

all

Returns all files from the given directory

clean_name

Sanitizes the given target/env name and returns it, but will exit the process if it does not uniquely identify a single target/env file. Set the last arg to true to return the path to the matching target instead.

clear

delete_val

Deletes the given key (and its val) from .origen/application.toml if it exists

delete_val_at_root

get

Gets the currently enabled targets

matches

Returns an array of possible target/environment files that match the given name/snippet

remove

Disables currently enables targets in the workspace

reset

Resets (deletes) the target back to its default value

set

Sets the targets, overriding any that may be present

set_at_root

set_workspace

Sets the given key and value (currently only a string is supported) in .origen/application.toml

set_workspace_array

Sets an Array-of-Strings workspace variable

set_workspace_array_at_root