origen::core

Module config

Collects the main Origen configuration options from all origen.toml files found in the application and Origen CLI installation file system paths

Examples

use origen::ORIGEN_CONFIG;

println!("Server: {}", &ORIGEN_CONFIG.pkg_server);  // => "Server: https://pkgs.company.net:9292"

Structs

AuxillaryCommandsTOML

Config

ConfigMetadata

DatasetConfig

DefaultUserConfig

InitialUserConfig

LDAPConfig

PluginConfig

PluginsConfig