origen::core::frontend

Trait ReleaseScribe

pub trait ReleaseScribe {
    // Required methods
    fn release_note_file(&self) -> Result<PathBuf>;
    fn get_release_note(&self) -> Result<String>;
    fn get_release_note_from_file(&self) -> Result<String>;
    fn get_release_title(&self) -> Result<Option<String>>;
    fn history_tracking_file(&self) -> Result<PathBuf>;
    fn append_history(
        &self,
        version: &Version,
        title: Option<&str>,
        text: &str,
        dry_run: bool,
    ) -> Result<()>;

    // Provided methods
    fn rc_files(&self) -> Result<Vec<PathBuf>> { ... }
    fn publish(
        &self,
        release: &Version,
        title: Option<Option<&str>>,
        body: Option<&str>,
        dry_run: bool,
    ) -> Result<Vec<PathBuf>> { ... }
}

Required Methods

fn release_note_file(&self) -> Result<PathBuf>

Returns the path to where a release note could be. The release note need not exists at this point, but this is where the release scribe will expect it to be.

fn get_release_note(&self) -> Result<String>

Gets the release note, either from an existing file or from a release dialogue.

fn get_release_note_from_file(&self) -> Result<String>

Gets the release note from an existing file or returns an error.

fn get_release_title(&self) -> Result<Option<String>>

Starts dialogue to retrieve a release title

fn history_tracking_file(&self) -> Result<PathBuf>

Returns the path to the history tracking file. This need not exists at this point, but is where it will be expected/created.

fn append_history( &self, version: &Version, title: Option<&str>, text: &str, dry_run: bool, ) -> Result<()>

Updates the history file given the release, title, and release note body

Provided Methods

fn rc_files(&self) -> Result<Vec<PathBuf>>

Returns a list of all files that should be checked into revision control. By default, this is only the history toml file.

fn publish( &self, release: &Version, title: Option<Option<&str>>, body: Option<&str>, dry_run: bool, ) -> Result<Vec<PathBuf>>

Goes through the process of getting release notes, updating the files, and returning whatever should be checked into revision control. Title and body can be given if derived elsewhere (e.g., from CLI), but normal retrieval process ensues if empty.