Module origen::core::reference_files

This manages how generated files are saved as references. When new or changed files are generated a save ref is saved on disk which will allow a future execution of the 'origen save_ref' command to determine the source and destination of the file to be saved as a the new reference version. Save refs are stored as single files per reference instead of a single file containing all references or the log as was used in O1. A single file is not used to avoid contentions when multiple generation jobs are running in parallel on the LSF.

Functions

apply_all_changed_refs

apply_all_new_refs

apply_ref

clear_save_refs

create_changed_ref

create_new_ref