origen::core::status

Struct Status

pub struct Status {
    pub is_origen_present: bool,
    pub is_app_present: bool,
    pub app: Option<Application>,
    pub origen_wksp_root: PathBuf,
    pub origen_version: Version,
    pub start_time: Tm,
    pub home: PathBuf,
    pub log_level: u8,
    pub is_app_in_origen_dev_mode: bool,
    pub origen_core_support_version: Version,
    pub origen_metal_backend_version: Version,
    /* private fields */
}

Exposes some status information about the runtime environment, e.g. whether an application workspace is present

Fields
is_origen_present: bool

When true, Origen is executing within an origen development workspace

is_app_present: bool

When true, Origen is executing within an application workspace

app: Option<Application>

When Origen is executing with the context of an application, this represents it

origen_wksp_root: PathBuf

When Origen is running within an Origen development workspace or when it is running within an app which is referencing a local copy of Origen, this will point to the root of the Origen workspace. i.e. it is only valid when either is_origen_present or is_app_in_origen_dev_mode is true.

origen_version: Version

The Origen version in a Semver object

start_time: Tmhome: PathBuf

The full file system path to the user’s home directory

log_level: u8is_app_in_origen_dev_mode: bool

When true it means that Origen is running within an app and that app is using a local development version of Origen

origen_core_support_version: Versionorigen_metal_backend_version: Version
Implementations

impl Status

pub fn register_custom_tester(&self, name: &str)

pub fn custom_tester_ids(&self) -> Vec<String>

pub fn generate_unique_id(&self) -> usize

Returns an ID that it guaranteed unique across threads and within the lifetime of an Origen invocation

pub fn push_testers_eq(&self, testers: Vec<SupportedTester>) -> OrigenResult<()>

Corresponds to the start of a tester specific block of code, returns an error if the given tester selection is not a subset of any existing selection

pub fn push_testers_neq( &self, testers: Vec<SupportedTester>, ) -> OrigenResult<()>

Corresponds to the start of a tester except (tester not equal to) block of code, returns an error if the given tester selection is not valid within any existing eq or neq blocks

pub fn pop_testers_eq(&self) -> OrigenResult<()>

Corresponds to the end of a tester specific block, returns an error if no tester specific selection currently exists

pub fn pop_testers_neq(&self) -> OrigenResult<()>

Corresponds to the end of a tester exclusion block, returns an error if no tester exlcusion selection currently exists

pub fn unhandled_error_count(&self) -> usize

Returns the number of unhandled errors that have been encountered since this thread started. An example of a unhandled error is a pattern that failed to generate. If an error occurs on the Python side then Origen will most likely crash, however on the rust side it is possible to keep going (e.g. to the next pattern), and this keeps track of how many big problems there were for reporting to the user at the end.

pub fn inc_unhandled_error_count(&self)

pub fn set_debug_enabled(&self, val: bool)

pub fn is_debug_enabled(&self) -> bool

pub fn set_operation(&self, val: Operation)

pub fn operation(&self) -> Operation

pub fn set_cli_location(&self, loc: Option<String>)

pub fn set_cli_version(&self, ver: Option<String>)

pub fn cli_location(&self) -> Option<PathBuf>

pub fn cli_version(&self) -> Option<Version>

pub fn set_fe_pkg_loc(&self, loc: Option<PathBuf>)

pub fn fe_pkg_loc(&self) -> RwLockReadGuard<'_, Option<PathBuf>>

pub fn set_fe_exe_loc(&self, loc: Option<PathBuf>)

pub fn fe_exe_loc(&self) -> RwLockReadGuard<'_, Option<PathBuf>>

pub fn update_other_build_info(&self, key: &str, item: &str) -> OrigenResult<()>

pub fn other_build_info(&self) -> HashMap<String, String>

pub fn set_output_dir(&self, path: &Path)

Set the base output dir to the given path, it is <APP ROOT>/output by default

pub fn set_reference_dir(&self, path: &Path)

Set the base reference dir to the given path, it is <APP ROOT>/.ref by default

pub fn in_origen_core_app(&self) -> bool

pub fn set_in_origen_core_app(&self, stat: bool)

pub fn set_dependency_src(&self, src: Option<DependencySrc>)

pub fn dependency_src(&self) -> RwLockReadGuard<'_, Option<DependencySrc>>

pub fn output_dir(&self) -> PathBuf

This is the main method to get the current output directory, accounting for all possible ways to set it, from current command, the app, default, etc. If nothing has been set (only possible when running globally), then it will default to the PWD. It will ensure that the directory exists before returning it.

pub fn with_output_dir<T, F>(&self, change_dir: bool, f: F) -> OrigenResult<T>
where F: FnMut(&Path) -> OrigenResult<T>,

Execute the given function with a reference to the current output directory (<APP ROOT>/output by default). Optionally, the current working directory can be switched to the output dir before executing the function and then restored at the end by setting change_to to True. If this is called when Origen is executing outside of an application workspace then it will return None unless it has been previously set by calling set_output_dir().

pub fn reference_dir(&self) -> Option<PathBuf>

This is the main method to get the current reference directory, accounting for all possible ways to set it, from current command, the app, default, etc. If nothing has been set (only possible when running globally), then it will return None. It will ensure that the directory exists before returning it.

pub fn with_reference_dir<T, F>( &self, change_dir: bool, f: F, ) -> OrigenResult<T>
where F: FnMut(Option<&PathBuf>) -> OrigenResult<T>,

Execute the given function with a reference to the current reference directory (<APP ROOT>/.ref by default). Optionally, the current working directory can be switched to the reference dir before executing the function and then restored at the end by setting change_to to True. If this is called when Origen is executing outside of an application workspace then it will return None unless it has been previously set by calling set_reference_dir().