origen::core::tester

Struct Tester

pub struct Tester {
    pub target_testers: Vec<TesterSource>,
    pub stats: Stats,
    /* private fields */
}

Fields
target_testers: Vec<TesterSource>

This is the testers that have been selected by the current target, it will be cleared and new testers will be pushed to it during a target load. It contains references to both Rust and Python domain testers.

stats: Stats

Keeps track of some stats, like how many patterns have been generated, how many with diffs, etc.

Implementations

impl Tester

pub fn new() -> Self

pub fn start_tester_eq_block( &self, testers: Vec<SupportedTester>, ) -> Result<(usize, usize)>

Starts a new tester-specific section in the current pattern and/or test program. The returned ID should be kept and given to end_tester_eq_block when the tester specific section is complete.

pub fn end_tester_eq_block( &self, pat_ref_id: usize, prog_ref_id: usize, ) -> Result<()>

Ends an open tester-specific section in the current pattern and/or test program. The ID produced when opening the block (via start_tester_eq_block) should be supplied as the main argument.

pub fn start_tester_neq_block( &self, testers: Vec<SupportedTester>, ) -> Result<(usize, usize)>

Like start_tester_eq_block, but the contained block will be included for all testers except those in the given list

pub fn end_tester_neq_block( &self, pat_ref_id: usize, prog_ref_id: usize, ) -> Result<()>

pub fn custom_tester_ids(&self) -> Vec<String>

pub fn _current_timeset_id(&self) -> Result<usize>

pub fn _current_pin_header_id(&self) -> Result<usize>

pub fn init(&mut self) -> Result<()>

pub fn reset(&mut self) -> Result<()>

This will be called by Origen immediately before it is about to load the target, it unloads all tester targets and all other state making it ready to accept a new set of targets

pub fn register_external_tester( &mut self, tester: &str, ) -> Result<SupportedTester>

pub fn get_timeset(&self, dut: &Dut) -> Option<Timeset>

pub fn _get_timeset(&self, dut: &Dut) -> Result<Timeset>

pub fn set_timeset( &mut self, dut: &Dut, model_id: usize, timeset_name: &str, ) -> Result<()>

pub fn clear_timeset(&mut self) -> Result<()>

pub fn get_pin_header(&self, dut: &Dut) -> Option<PinHeader>

pub fn _get_pin_header(&self, dut: &Dut) -> Result<PinHeader>

pub fn set_pin_header( &mut self, dut: &Dut, model_id: usize, pin_header_name: &str, ) -> Result<()>

pub fn clear_pin_header(&mut self) -> Result<()>

pub fn issue_callback_at(&mut self, idx: usize) -> Result<()>

pub fn cc(&mut self, comment: &str) -> Result<()>

pub fn cycle(&mut self, repeat: Option<usize>) -> Result<()>

pub fn overlay(&self, overlay: &Overlay) -> Result<()>

pub fn capture(&self, capture: &Capture) -> Result<()>

pub fn generate_pattern_header( &self, app_comments: Option<Vec<String>>, pattern_comments: Option<Vec<String>>, ) -> Result<()>

pub fn end_pattern(&self) -> Result<()>

pub fn render_pattern_for_target_at( &mut self, idx: usize, diff_and_display: bool, ) -> Result<Vec<PathBuf>>

Renders the output for the target at index i. Allows the frontend to call testers in a loop. When diff_and_display is true the generated files will be displayed to the console and checked for diffs.

pub fn render_program_for_target_at( &mut self, idx: usize, diff_and_display: bool, ) -> Result<(Vec<PathBuf>, Model)>

Renders the test program for the target at index i. Allows the frontend to call testers in a (multithreaded) loop. When diff_and_display is true the generated files will be displayed to the console and checked for diffs.

pub fn target(&mut self, tester: SupportedTester) -> Result<&TesterSource>

pub fn targets(&self) -> &Vec<TesterSource>

pub fn targets_as_strs(&self) -> Vec<String>

pub fn focused_tester(&self) -> Option<&TesterSource>

pub fn focused_tester_name(&self) -> Option<String>

pub fn prepare_for_generate(&mut self) -> Result<()>

This is called automatically at the very start of a generate command, it is invoked from Python, so state can be established here which will persist for the rest of the command