Enum origen::generator::ast::Attrs

pub enum Attrs {
    _Inline,
    T(usize),
    Integer(i64),
    Float(f64),
    String(String),
    TesterEq(Vec<SupportedTester>),
    TesterNeq(Vec<SupportedTester>),
    Test(String),
    Comment(u8, String),
    SetTimeset(usize),
    ClearTimeset,
    SetPinHeader(usize),
    ClearPinHeader,
    PinGroupAction(usize, Vec<String>, Option<HashMap<String, Metadata>>),
    PinAction(usize, String, Option<HashMap<String, Metadata>>),
    Opcode(String, IndexMap<String, String>),
    Cycle(u32, bool),
    PatternHeader,
    PatternEnd,
    RegWrite(Transaction),
    RegVerify(Transaction),
    JTAGWriteIR(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGVerifyIR(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGWriteDR(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGVerifyDR(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGReset(usize, Option<IndexMap<String, Metadata>>),
    SWDWriteAP(usize, Transaction, Acknowledgements, Option<IndexMap<String, Metadata>>),
    SWDVerifyAP(usize, Transaction, Acknowledgements, Option<bool>, Option<IndexMap<String, Metadata>>),
    SWDWriteDP(usize, Transaction, Acknowledgements, Option<IndexMap<String, Metadata>>),
    SWDVerifyDP(usize, Transaction, Acknowledgements, Option<bool>, Option<IndexMap<String, Metadata>>),
    SWDLineReset,
    JTAGDPWriteDP(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGDPVerifyDP(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGDPWriteAP(usize, Transaction, Option<IndexMap<String, Metadata>>),
    JTAGDPVerifyAP(usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugMemAPWriteReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugMemAPWriteInternalReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugMemAPVerifyReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugMemAPVerifyInternalReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugWriteDP(usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugVerifyDP(usize, Transaction, Option<IndexMap<String, Metadata>>),
    ArmDebugSwjJTAGToSWD(usize),
    ArmDebugSwjSWDToJTAG(usize),
    TextSection(Option<String>, Option<u8>),
    TextBoundaryLine,
    TextLine,
    Text(String),
    User,
    OrigenCommand(String),
    Timestamp,
    Mode,
    TargetsStacked,
    OS,
    AppRoot,
    OrigenVersion,
    OrigenRoot,
    PGMFlow(String),
    PGMSubFlow(String),
    PGMDefTest(usize, String, SupportedTester, String, String),
    PGMDefTestInv(usize, String, SupportedTester),
    PGMAssignTestToInv(usize, usize),
    PGMSetAttr(usize, String, ParamValue),
    PGMTest(usize, FlowID),
    PGMTestStr(String, FlowID),
    PGMPatternGroup(usize, String, SupportedTester, Option<PatternGroupType>),
    PGMPushPattern(usize, String, Option<String>),
    PGMRender(String),
    PGMLog(String),
    PGMGroup(String, Option<SupportedTester>, GroupType, Option<FlowID>),
    PGMCondition(FlowCondition),
    PGMCz(usize, String, FlowID),
    PGMDefBin(usize, bool, BinType, Option<String>, Option<usize>),
    PGMBin(usize, Option<usize>, BinType),
    PGMOnFailed(FlowID),
    PGMOnPassed(FlowID),
    PGMIGXLSetWaitFlags(usize, Vec<String>),
    STIL,
    STILUnknown,
    STILVersion(u32, u32),
    STILHeader,
    STILTitle(String),
    STILDate(String),
    STILSource(String),
    STILHistory,
    STILAnnotation(String),
    STILInclude(String, Option<String>),
    STILSignals,
    STILSignal(String, SignalType),
    STILTermination(Termination),
    STILDefaultState(State),
    STILBase(Base, String),
    STILAlignment(Alignment),
    STILScanIn(u32),
    STILScanOut(u32),
    STILDataBitCount(u32),
    STILSignalGroups(Option<String>),
    STILSignalGroup(String),
    STILSigRefExpr,
    STILTimeExpr,
    STILSIUnit(String),
    STILEngPrefix(String),
    STILAdd,
    STILSubtract,
    STILMultiply,
    STILDivide,
    STILParens,
    STILNumberWithUnit,
    STILPatternExec(Option<String>),
    STILCategoryRef(String),
    STILSelectorRef(String),
    STILTimingRef(String),
    STILPatternBurstRef(String),
    STILPatternBurst(String),
    STILSignalGroupsRef(String),
    STILMacroDefs(String),
    STILProcedures(String),
    STILScanStructuresRef(String),
    STILStart(String),
    STILStop(String),
    STILTerminations,
    STILTerminationItem,
    STILPatList,
    STILPat(String),
    STILLabel(String),
    STILTiming(Option<String>),
    STILWaveformTable(String),
    STILPeriod,
    STILInherit(String),
    STILSubWaveforms,
    STILSubWaveform,
    STILWaveforms,
    STILWaveform,
    STILWFChar(String),
    STILEvent,
    STILEventList(Vec<char>),
    STILSpec(Option<String>),
    STILCategory(String),
    STILSpecItem,
    STILTypicalVar(String),
    STILSpecVar(String),
    STILSpecVarItem(Selector),
    STILVariable(String),
    STILSelector(String),
    STILSelectorItem(String, Selector),
    STILScanStructures(Option<String>),
    STILScanChain(String),
    STILScanLength(u64),
    STILScanOutLength(u64),
    STILScanCells,
    STILScanMasterClock,
    STILScanSlaveClock,
    STILScanInversion(u8),
    STILScanInName(String),
    STILScanOutName(String),
    STILNot,
    STILPattern(String),
    STILTimeUnit,
    STILVector,
    STILCyclizedData,
    STILNonCyclizedData,
    STILRepeat(u64),
    STILWaveformFormat,
    STILHexFormat(Option<String>),
    STILDecFormat(Option<String>),
    STILData(String),
    STILTimeValue(u64),
    STILWaveformRef(String),
    STILCondition,
    STILCall(String),
    STILMacro(String),
    STILLoop(u64),
    STILMatchLoop(Option<u64>),
    STILGoto(String),
    STILBreakPoint,
    STILIDDQ,
    STILStopStatement,
}

Variants

_Inline

T(usize)

Integer(i64)

Float(f64)

String(String)

TesterEq(Vec<SupportedTester>)

TesterNeq(Vec<SupportedTester>)

Test(String)

Comment(u8, String)

SetTimeset(usize)

ClearTimeset

SetPinHeader(usize)

ClearPinHeader

PinGroupAction(usize, Vec<String>, Option<HashMap<String, Metadata>>)

PinAction(usize, String, Option<HashMap<String, Metadata>>)

Opcode(String, IndexMap<String, String>)

Cycle(u32, bool)

PatternHeader

PatternEnd

RegWrite(Transaction)

RegVerify(Transaction)

JTAGWriteIR(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGVerifyIR(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGWriteDR(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGVerifyDR(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGReset(usize, Option<IndexMap<String, Metadata>>)

SWDWriteAP(usize, Transaction, Acknowledgements, Option<IndexMap<String, Metadata>>)

SWDVerifyAP(usize, Transaction, Acknowledgements, Option<bool>, Option<IndexMap<String, Metadata>>)

SWDWriteDP(usize, Transaction, Acknowledgements, Option<IndexMap<String, Metadata>>)

SWDVerifyDP(usize, Transaction, Acknowledgements, Option<bool>, Option<IndexMap<String, Metadata>>)

SWDLineReset

JTAGDPWriteDP(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGDPVerifyDP(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGDPWriteAP(usize, Transaction, Option<IndexMap<String, Metadata>>)

JTAGDPVerifyAP(usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugMemAPWriteReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugMemAPWriteInternalReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugMemAPVerifyReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugMemAPVerifyInternalReg(usize, usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugWriteDP(usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugVerifyDP(usize, Transaction, Option<IndexMap<String, Metadata>>)

ArmDebugSwjJTAGToSWD(usize)

ArmDebugSwjSWDToJTAG(usize)

TextSection(Option<String>, Option<u8>)

TextBoundaryLine

TextLine

Text(String)

User

OrigenCommand(String)

Timestamp

Mode

TargetsStacked

OS

AppRoot

OrigenVersion

OrigenRoot

PGMFlow(String)

PGMSubFlow(String)

PGMDefTest(usize, String, SupportedTester, String, String)

Defines a new test, this must be done before attributes can be set via PGMSetAttr. Note that this doesn't actually add it to the test flow, that must be done via a PGMTest node ID, Name, Tester, Library Template

PGMDefTestInv(usize, String, SupportedTester)

Defines a new test invocation, this must be done before attributes can be set via PGMSetAttr. Note that this doesn't actually add it to the test flow, that must be done via a PGMTest node ID, Name, Tester

PGMAssignTestToInv(usize, usize)

Assign an existing test to an existing invocation InvID TestID

PGMSetAttr(usize, String, ParamValue)

Set the attribute with the given name within the given test (ID), to the given value

PGMTest(usize, FlowID)

Execute a test (or invocation) from the flow

PGMTestStr(String, FlowID)

Execute a test (or invocation) from the flow, where the test is simply a string to be inserted into the flow

PGMPatternGroup(usize, String, SupportedTester, Option<PatternGroupType>)

Defines a new pattern group, also used to model IG-XL pattern sets

PGMPushPattern(usize, String, Option<String>)

Push a pattern to the given pattern group ID

PGMRender(String)

Render the given text directly to the flow

PGMLog(String)

Add a log line to the flow

PGMGroup(String, Option<SupportedTester>, GroupType, Option<FlowID>)

A FlowID will always be present when the group type is a flow group

PGMCondition(FlowCondition)

All children will be gated by the given condition (if_failed, if_enabled, etc.)

PGMCz(usize, String, FlowID)

Execute a test (or invocation) from the flow with a CZ setup reference

PGMDefBin(usize, bool, BinType, Option<String>, Option<usize>)

Bin (number, is_soft, type, description, priority)

PGMBin(usize, Option<usize>, BinType)

Bin out (hard, soft, type)

PGMOnFailed(FlowID)

Events to run if the test or group with the given ID failed

PGMOnPassed(FlowID)

Events to run if the test or group with the given ID passed

PGMIGXLSetWaitFlags(usize, Vec<String>)

STIL

STILUnknown

STILVersion(u32, u32)

STILHeader

STILTitle(String)

STILDate(String)

STILSource(String)

STILHistory

STILAnnotation(String)

STILInclude(String, Option<String>)

STILSignals

STILSignal(String, SignalType)

STILTermination(Termination)

STILDefaultState(State)

STILBase(Base, String)

STILAlignment(Alignment)

STILScanIn(u32)

STILScanOut(u32)

STILDataBitCount(u32)

STILSignalGroups(Option<String>)

STILSignalGroup(String)

STILSigRefExpr

STILTimeExpr

STILSIUnit(String)

STILEngPrefix(String)

STILAdd

STILSubtract

STILMultiply

STILDivide

STILParens

STILNumberWithUnit

STILPatternExec(Option<String>)

STILCategoryRef(String)

STILSelectorRef(String)

STILTimingRef(String)

STILPatternBurstRef(String)

STILPatternBurst(String)

STILSignalGroupsRef(String)

STILMacroDefs(String)

STILProcedures(String)

STILScanStructuresRef(String)

STILStart(String)

STILStop(String)

STILTerminations

STILTerminationItem

STILPatList

STILPat(String)

STILLabel(String)

STILTiming(Option<String>)

STILWaveformTable(String)

STILPeriod

STILInherit(String)

STILSubWaveforms

STILSubWaveform

STILWaveforms

STILWaveform

STILWFChar(String)

STILEvent

STILEventList(Vec<char>)

STILSpec(Option<String>)

STILCategory(String)

STILSpecItem

STILTypicalVar(String)

STILSpecVar(String)

STILSpecVarItem(Selector)

STILVariable(String)

STILSelector(String)

STILSelectorItem(String, Selector)

STILScanStructures(Option<String>)

STILScanChain(String)

STILScanLength(u64)

STILScanOutLength(u64)

STILScanCells

STILScanMasterClock

STILScanSlaveClock

STILScanInversion(u8)

STILScanInName(String)

STILScanOutName(String)

STILNot

STILPattern(String)

STILTimeUnit

STILVector

STILCyclizedData

STILNonCyclizedData

STILRepeat(u64)

STILWaveformFormat

STILHexFormat(Option<String>)

STILDecFormat(Option<String>)

STILData(String)

STILTimeValue(u64)

STILWaveformRef(String)

STILCondition

STILCall(String)

STILMacro(String)

STILLoop(u64)

STILMatchLoop(Option<u64>)

STILGoto(String)

STILBreakPoint

STILIDDQ

STILStopStatement