Struct origen::generator::ast::AST

pub struct AST { /* fields omitted */ }

An AST provides an API for constructing a node tree, when completed it can be unwrapped to a node by calling the unwrap() method

Implementations

impl AST

pub fn new() -> AST

Create a new AST with the given node as the top-level

pub fn unwrap(&mut self) -> Node

Consumes the AST, converting it to a Node

pub fn push(&mut self, node: Node)

Push a new terminal node into the AST

pub fn push_and_open(&mut self, node: Node) -> usize

Push a new node into the AST and leave it open, meaning that all new nodes added to the AST will be inserted as children of this node until it is closed. A reference ID is returned and the caller should save this and provide it again when calling close(). If the reference does not match the expected an error will be raised. This will catch any cases of AST application code forgetting to close a node before closing one of its parents.

pub fn close(&mut self, ref_id: usize) -> Result<()>

Close the currently open node

pub fn replace(&mut self, node: Node, offset: usize) -> Result<()>

Replace the node n - offset with the given node, use offset = 0 to replace the last node that was pushed. Fails if the AST has no children yet or if the offset is otherwise out of range.

pub fn insert(&mut self, node: Node, offset: usize) -> Result<()>

Insert the node at position n - offset, using offset = 0 is equivalent calling push().

pub fn get(&self, offset: usize) -> Result<Node>

Returns a copy of node n - offset, an offset of 0 means the last node pushed. Fails if the offset is out of range.

pub fn start(&mut self, node: Node)

Clear the current AST and start a new one with the given node at the top-level

pub fn process(&self, process_fn: &mut dyn FnMut(&Node) -> Node) -> Node

pub fn with_node<T, F>(&self, process_fn: F) -> Result<T> where
F: FnMut(&Node) -> Result<T>,

Execute the given function which receives the a reference to the AST (as a Node) as its input, returning the result of the function

pub fn to_string(&self) -> String

pub fn to_pickle(&self) -> Vec<u8>

Serializes the AST for import into Python

pub fn to_node(&self) -> Node

Clones the current state of the AST into a Node, leaving the AST unmodified