Struct origen::generator::ast::Node

pub struct Node {
    pub attrs: Attrs,
    pub meta: Option<Meta>,
    pub children: Vec<Box<Node>>,
}

Fields
attrs: Attrsmeta: Option<Meta>children: Vec<Box<Node>>
Implementations

impl Node

pub fn new(attrs: Attrs) -> Node

pub fn unwrap(&mut self) -> Option<Node>

pub fn process(&self, processor: &mut dyn Processor) -> Result<Option<Node>>

Returns a new node which is the output of the node processed by the given processor. Returning None means that the processor has decided that the node should be removed from the next stage AST.

pub fn new_with_meta(
attrs: Attrs,
filename: Option<String>,
lineno: Option<usize>
) -> Node

pub fn to_string(&self) -> String

pub fn to_pickle(&self) -> Vec<u8>

Serializes the AST for import into Python

pub fn add_child(&mut self, node: Node)

pub fn add_children(&mut self, nodes: Vec<Node>) -> &Self

pub fn insert_child(&mut self, node: Node, offset: usize) -> Result<()>

pub fn replace_child(&mut self, node: Node, offset: usize) -> Result<()>

Replace the child n - offset with the given node, use offset = 0 to replace the last child that was pushed. Fails if the node has no children or if the given offset is otherwise out of range.

pub fn get_child(&self, offset: usize) -> Result<Node>

Returns a copy of child n - offset, an offset of 0 means the last child that was pushed. Fails if the node has no children or if the given offset is otherwise out of range.

pub fn process_return_code(
&self,
code: Return,
processor: &mut dyn Processor
) -> Result<Option<Node>>

pub fn process_and_update_children(
&self,
processor: &mut dyn Processor
) -> Result<Node>

Returns a new node which is a copy of self with its children replaced by their processed counterparts.

pub fn process_and_box_children(
&self,
processor: &mut dyn Processor
) -> Result<Vec<Box<Node>>>

Returns processed versions of the node's children, each wrapped in a Box

pub fn process_children(
&self,
processor: &mut dyn Processor
) -> Result<Vec<Node>>

Returns processed versions of the node's children

pub fn replace_children(&self, nodes: Vec<Box<Node>>) -> Node

Returns a new node which is a copy of self with its children replaced by the given collection of nodes.

pub fn replace_attrs(&self, attrs: Attrs) -> Node

Returns a new node which is a copy of self with its attrs replaced by the given attrs.