origen::generator

Enum PAT

pub enum PAT {

    T(usize),
    Nil,
    Integer(i64),
    Float(f64),
    String(String),
    TesterEq(Vec<SupportedTester>),
    TesterNeq(Vec<SupportedTester>),
    Test(String),
    Comment(u8, String),
    SetTimeset(usize),
    ClearTimeset,
    SetPinHeader(usize),
    ClearPinHeader,
    PinGroupAction(usize, Vec<String>, Option<TypedValueMap>),
    PinAction(usize, String, Option<TypedValueMap>),
    Capture(Capture, Option<TypedValueMap>),
    EndCapture(Option<usize>),
    Overlay(Overlay, Option<TypedValueMap>),
    EndOverlay(Option<String>, Option<usize>),
    Opcode(String, IndexMap<String, String>),
    Cycle(u32, bool),
    PatternHeader,
    PatternEnd,
    RegWrite(Transaction),
    RegVerify(Transaction),
    RegCapture(Transaction),
    RegOverlay(Transaction),
    JTAGWriteIR(usize, Transaction, Option<TypedValueMap>),
    JTAGVerifyIR(usize, Transaction, Option<TypedValueMap>),
    JTAGWriteDR(usize, Transaction, Option<TypedValueMap>),
    JTAGVerifyDR(usize, Transaction, Option<TypedValueMap>),
    JTAGReset(usize, Option<TypedValueMap>),
    SWDWriteAP(usize, Transaction, Acknowledgements, Option<TypedValueMap>),
    SWDVerifyAP(usize, Transaction, Acknowledgements, Option<bool>, Option<TypedValueMap>),
    SWDWriteDP(usize, Transaction, Acknowledgements, Option<TypedValueMap>),
    SWDVerifyDP(usize, Transaction, Acknowledgements, Option<bool>, Option<TypedValueMap>),
    SWDLineReset,
    JTAGDPWriteDP(usize, Transaction, Option<TypedValueMap>),
    JTAGDPVerifyDP(usize, Transaction, Option<TypedValueMap>),
    JTAGDPWriteAP(usize, Transaction, Option<TypedValueMap>),
    JTAGDPVerifyAP(usize, Transaction, Option<TypedValueMap>),
    ArmDebugMemAPWriteReg(usize, usize, Transaction, Option<TypedValueMap>),
    ArmDebugMemAPWriteInternalReg(usize, usize, Transaction, Option<TypedValueMap>),
    ArmDebugMemAPVerifyReg(usize, usize, Transaction, Option<TypedValueMap>),
    ArmDebugMemAPVerifyInternalReg(usize, usize, Transaction, Option<TypedValueMap>),
    ArmDebugWriteDP(usize, Transaction, Option<TypedValueMap>),
    ArmDebugVerifyDP(usize, Transaction, Option<TypedValueMap>),
    ArmDebugSwjJTAGToSWD(usize),
    ArmDebugSwjSWDToJTAG(usize),
    SimpleProtocolReset(usize),
    SimpleProtocolWrite(usize, Transaction),
    SimpleProtocolVerify(usize, Transaction),
    TextSection(Option<String>, Option<u8>),
    TextBoundaryLine,
    TextLine,
    Text(String),
    User,
    OrigenCommand(String),
    Timestamp,
    Mode,
    TargetsStacked,
    OS,
    AppRoot,
    OrigenVersion,
    OrigenRoot,

}

Variants

T(usize)

A simple node that is quick to write and use in processor unit tests

Nil

This will be ignored by all processors, so can be used to indicate the absence of a node

Integer(i64)

Float(f64)

String(String)

TesterEq(Vec<SupportedTester>)

TesterNeq(Vec<SupportedTester>)

Test(String)

Comment(u8, String)

SetTimeset(usize)

ClearTimeset

SetPinHeader(usize)

ClearPinHeader

PinGroupAction(usize, Vec<String>, Option<TypedValueMap>)

PinAction(usize, String, Option<TypedValueMap>)

Capture(Capture, Option<TypedValueMap>)

EndCapture(Option<usize>)

Overlay(Overlay, Option<TypedValueMap>)

EndOverlay(Option<String>, Option<usize>)

Opcode(String, IndexMap<String, String>)

Cycle(u32, bool)

PatternHeader

PatternEnd

RegWrite(Transaction)

RegVerify(Transaction)

RegCapture(Transaction)

RegOverlay(Transaction)

JTAGWriteIR(usize, Transaction, Option<TypedValueMap>)

JTAGVerifyIR(usize, Transaction, Option<TypedValueMap>)

JTAGWriteDR(usize, Transaction, Option<TypedValueMap>)

JTAGVerifyDR(usize, Transaction, Option<TypedValueMap>)

JTAGReset(usize, Option<TypedValueMap>)

SWDWriteAP(usize, Transaction, Acknowledgements, Option<TypedValueMap>)

SWDVerifyAP(usize, Transaction, Acknowledgements, Option<bool>, Option<TypedValueMap>)

SWDWriteDP(usize, Transaction, Acknowledgements, Option<TypedValueMap>)

SWDVerifyDP(usize, Transaction, Acknowledgements, Option<bool>, Option<TypedValueMap>)

SWDLineReset

JTAGDPWriteDP(usize, Transaction, Option<TypedValueMap>)

JTAGDPVerifyDP(usize, Transaction, Option<TypedValueMap>)

JTAGDPWriteAP(usize, Transaction, Option<TypedValueMap>)

JTAGDPVerifyAP(usize, Transaction, Option<TypedValueMap>)

ArmDebugMemAPWriteReg(usize, usize, Transaction, Option<TypedValueMap>)

ArmDebugMemAPWriteInternalReg(usize, usize, Transaction, Option<TypedValueMap>)

ArmDebugMemAPVerifyReg(usize, usize, Transaction, Option<TypedValueMap>)

ArmDebugMemAPVerifyInternalReg(usize, usize, Transaction, Option<TypedValueMap>)

ArmDebugWriteDP(usize, Transaction, Option<TypedValueMap>)

ArmDebugVerifyDP(usize, Transaction, Option<TypedValueMap>)

ArmDebugSwjJTAGToSWD(usize)

ArmDebugSwjSWDToJTAG(usize)

SimpleProtocolReset(usize)

SimpleProtocolWrite(usize, Transaction)

SimpleProtocolVerify(usize, Transaction)

TextSection(Option<String>, Option<u8>)

TextBoundaryLine

TextLine

Text(String)

User

OrigenCommand(String)

Timestamp

Mode

TargetsStacked

OS

AppRoot

OrigenVersion

OrigenRoot