Enum origen::generator::processor::Return

pub enum Return {
    None,
    Unmodified,
    ProcessChildren,
    Replace(Node),
    Unwrap,
    Inline(Vec<Node>),
    InlineBoxed(Vec<Box<Node>>),
    UnwrapWithProcessedChildren,
    InlineWithProcessedChildren(Vec<Node>),
}

All procesor handler methods should return this

Variants

None

Deletes the node from the output AST.

Unmodified

Clones the node (and all of its children) into the output AST. Note that the child nodes are not processed in this case (though they will appear in the output unmodified).

ProcessChildren

Clones the node but replaces it's current children with their processed counterparts in the output AST.

Replace(Node)

Replace the node in the output AST with the given node.

Unwrap

Removes the node and leaves its children in its place.

Inline(Vec<Node>)

Replace the node in the output AST with the given nodes, the vector wrapper will be removed and the nodes will be placed inline with where the current node is/was.

InlineBoxed(Vec<Box<Node>>)

Same as Inline, but accepts a vector of boxed nodes

UnwrapWithProcessedChildren

A combination of Unwrap and ProcessChildren, which will unwrap the current node but leave processed children in its place

InlineWithProcessedChildren(Vec<Node>)

A combinatin of Inline and ProcessChildren which will add the given nodes then proceed to process the original node's children