origen::generator::utility::transaction

Struct Transaction

pub struct Transaction {
    pub action: Option<Action>,
    pub reg_id: Option<usize>,
    pub address: Option<BigUint>,
    pub address_width: Option<usize>,
    pub width: usize,
    pub data: BigUint,
    pub bit_enable: BigUint,
    pub capture: Option<Capture>,
    pub overlay: Option<Overlay>,
    pub set_actions: Option<Vec<PinAction>>,
    pub metadata: Option<TypedValueMap>,
}

Fields
action: Option<Action>reg_id: Option<usize>address: Option<BigUint>address_width: Option<usize>width: usizedata: BigUintbit_enable: BigUintcapture: Option<Capture>overlay: Option<Overlay>set_actions: Option<Vec<PinAction>>metadata: Option<TypedValueMap>
Implementations

impl Transaction

pub fn new_write(data: BigUint, width: usize) -> Result<Self>

pub fn new_write_with_addr( data: BigUint, width: usize, addr: u128, ) -> Result<Self>

pub fn new_verify(data: BigUint, width: usize) -> Result<Self>

pub fn new_capture( width: usize, capture_enables: Option<BigUint>, ) -> Result<Self>

pub fn set_capture_enables( &mut self, capture_enables: Option<BigUint>, ) -> Result<()>

pub fn new_highz(width: usize) -> Result<Self>

pub fn apply_overlay( &mut self, label: Option<String>, symbol: Option<String>, enables: Option<BigUint>, ) -> Result<()>

pub fn apply_overlay_pin_ids(&mut self, pin_ids: &Vec<usize>) -> Result<bool>

pub fn has_overlay(&self) -> bool

pub fn new_set(actions: &Vec<PinAction>) -> Result<Self>

pub fn is_set_action(&self) -> bool

pub fn addr(&self) -> Result<u128>

pub fn addr_width(&self) -> Result<usize>

pub fn to_symbols(&self) -> Result<Vec<(String, bool, bool)>>

pub fn enable_of_width(width: usize) -> Result<BigUint>

pub fn enable_width(&self) -> Result<BigUint>

Helper method to generate a mask which enables all bits in the transaction

pub fn check_size(data: &BigUint, width: usize) -> Result<()>

pub fn to_dummy(&self) -> Result<Self>

pub fn to_addr_trans(&self, default_addr_size: Option<usize>) -> Result<Self>

Shortcut function to generate a new transaction where the address of this transaction acts as the data of the new one. The address_width field will be the new transaction’s width. If this is not provided, it will be taken from the default_addr_size. If neither widths are provided, or if the address width exceeds the resolved width, an error is returned.

pub fn prepend_data(&mut self, data: BigUint, width: usize) -> Result<()>

pub fn as_write_node(&self) -> Result<Node<PAT>>

pub fn as_verify_node(&self) -> Result<Node<PAT>>

pub fn as_capture_node(&self) -> Result<Node<PAT>>

pub fn chunk_data(&self, chunk_width: usize) -> Result<Vec<BigUint>>

pub fn chunk_addr(&self, chunk_width: usize) -> Result<Vec<BigUint>>

pub fn is_capture(&self) -> bool

pub fn resize(&mut self, new_width: usize) -> Result<()>

Updates the width, returning an error if data doesn’t fit in the new width.