Module origen::prog_gen

Modules

advantest

flow_api

teradyne

Structs

Bin

FlowID
Used to uniquely identify a test in a flow

FlowManager

Test
This is an abstract data object which is used to model test instances on Teradyne platforms and both test methods and test suites on Advantest platforms. A test template is modelled as a Test where indirect = true, which means that it will never be rendered directly to a test program output file, however it can be referenced as a parent by direct tests which are to be rendered. Child tests can add additional prameters/aliases/defaults and/or inherit or override those from parent tests.

Enums

BinType

FlowCondition

GroupType

ParamType

ParamValue

PatternGroupType