Module origen::prog_gen::flow_api

Functions

assign_test_to_invocation

bin
Bin out the DUT

define_bin

define_pattern_group

define_test
Defines a new test in the AST, returning its ID. A test must be initially defined before attributes can be set on it. It won't actually appear in the test flow until it is added to if via add_test()

define_test_invocation
Defines a new test invocation in the AST, returning its ID. A test invoration must be initially defined before attributes can be set on it. It won't actually appear in the test flow until it is added to if via add_test()

end_block

execute_cz_test
Cz the given test (or invocation) from the current flow

execute_test
Execute the given test (or invocation) from the current flow

execute_test_str
Execute the given test (or invocation) from the current flow, where the test is a string that will be rendered verbatim to the flow - no linkage to an actual test object will be checked or inserted by Origen

log

push_pattern_to_group

render
Renders the given string directly to the test flow

set_test_attr
Set an attribute of either a test or a test invocation

set_wait_flags
[IGXL only] Set the given wait flags on the given test instance

start_condition

start_group
Used to model flow groups, IG-XL test instance groups, etc.

start_on_failed
Start an on-failed block (events to run if the given test or group failed), the returned reference should be retained and passed to end_block

start_on_passed
Start an on-passed block (events to run if the given test or group passed), the returned reference should be retained and passed to end_block

start_sub_flow
Start a sub-flow, the returned reference should be retained and passed to end_block