Function origen::prog_gen::flow_api::define_test_invocation

pub fn define_test_invocation(
    name: &str, 
    tester: &SupportedTester, 
    meta: Option<Meta>
) -> Result<usize>

Defines a new test invocation in the AST, returning its ID. A test invoration must be initially defined before attributes can be set on it. It won't actually appear in the test flow until it is added to if via add_test()