origen

Module python

Re-exports

pub use crate::_generated::python::MIN_PYTHON_VERSION;

Macros

pyproject_str 🔒

user_env_str 🔒

Structs

Config

NO_ORIGEN_BOOT_MODULE_ERROR

PYTHON_CONFIG

Constants

PYPROJECT 🔒

USER_ENV 🔒

Functions

add_origen_env

Adds any Origen-related environment settings to a command

extract_version 🔒

get_current_user_and_email

Attempts to get the username and email, utilizing the full Origen environment (site config, etc.)

get_version 🔒

Get the Python version from the given command

is_backend_origen_mod_missing_err

python_version

Returns the version of the Python that was discovered for this workspace.

resolve_pyproject

run

Execute the given Python code

run_and_capture_stdout

Execute the given Python code and capture its stdout

run_with_callbacks

Run silently with all STDOUT and STDERR handled by the given callback functions

uv_version

Returns the installed UV version.