Function origen::revision_control::git::config

pub fn config(attr_name: &str) -> Option<String>

Attempts to get the given attribute from the Git config. Not sure if it's possible to do this via libgit2, so this currently uses regular Git unlike most of this driver. If Git is not available, or any other issue, then it will silently fail and simply return None. e.g. call git::config("email") to get the user's email address.