origen

Struct STATUS

pub struct STATUS { /* private fields */ }

Provides status information derived from the runtime environment, e.g. if an app is present If an app is present then its Application struct is stored in here. Things like the current output and reference directories should be derived from here.