origen

Module utility

Modules

big_uint_helpers

Helpers for working with BigUnits

command_helpers

file_actions

A collection of utilities for making modifications to existing files, e.g. to add/remove lines from them, etc.

file_utils

github

location

num_helpers

release_scribe

sessions

Functions

resolve_dir_from_app_root

Resolves a directory path from the current application root. Accepts an optional ‘user_val’ and a default. The resulting directory will be resolved from:

status_to_bool