Function origen::utility::command_helpers::log_stdout_and_stderr

pub fn log_stdout_and_stderr(
    process: &mut Child, 
    stdout_callback: Option<&mut dyn FnMut(&str)>, 
    stderr_callback: Option<&mut dyn FnMut(&str)>
)

Log both stdout and stderr to the debug and error logs respectively, optionally calling a callback function for each line captured. If no callbacks are given then any captures lines will be sent to the debug and error logs respectively. This currently logs stderr after stdout, in future this will hopefully support concurrent logging of both streams.