origen::utility

Module file_actions

A collection of utilities for making modifications to existing files, e.g. to add/remove lines from them, etc.

Functions

append

Append the given text at the end of the given file. Returns an error if any error is encountered when trying to make the update, e.g. if the given file doesn’t exist.

contains

Returns Ok(true) if any line in the given file matches the given regex

insert_after

Insert the given text into the given file immediately after the first text matched by the given regular expression. Returns an error if any error is encountered when trying to make the update, e.g. if the given file doesn’t exist. Otherwise, it will return true if the text was inserted and false if not, i.e. when the given regular expression did not match any line in the file.

insert_after_all

Like insert_after, but this function will insert the given text after every match rather than just the first one

insert_before

Insert the given text into the given file immediately before the first text matched by the given regular expression. Returns an error if any error is encountered when trying to make the update, e.g. if the given file doesn’t exist. Otherwise, it will return true if the text was inserted and false if not, i.e. when the given regular expression did not match any line in the file.

insert_before_all

Like insert_before, but this function will insert the given text before every match rather than just the first one

prepend

Prepend the given text at the start of the given file. Returns an error if any error is encountered when trying to make the update, e.g. if the given file doesn’t exist.

remove_line

Remove the first line in the given file that matches the given regex. Returns true if the file was modified.

remove_line_all

Like remove_line, but removes all matching lines in the given file.

replace

Replaces the first occurrence of the given regex with the given text, note that the regex is evaluated on a per line basis are therefore cannot span multiple lines

replace_all

Like replace, but replaces occurences of the given regex on all lines in the file