origen::utility

Module file_utils

Functions

_copy

copy

Copy the given file or directory to the given directory, directories will be copied recursively.

copy_contents

Like copy, however if the source is a directory then its contents will be copied to the target directory, rather than copying the source folder itself

mv

Move a file or directory

to_relative_path

Returns the given abs path as a relative path. By default it will be made relative to the PWD, but an alternative path to make it relative to can be supplied. Can return an error if