Function origen::utility::file_utils::with_dir

pub fn with_dir<T, F>(path: &Path, f: F) -> Result<T> where
    F: FnMut() -> Result<T>, 

Temporarily sets the current dir to the given dir for the duration of the given function and then restores it at the end. An error will be returned if there is a problem switching to the given directory, e.g. if it doesn't exist, otherwise the result from the given function is returned.

Examples

use std::path::Path;
use origen::utility::file_utils::with_dir;

let result = with_dir(Path::new("path/to/some/dir"), || {
  // Do something in that dir
  Ok(())
});