Module origen::utility::logger

The Origen logger is implemented as a singleton and data is logged via globally available macros.

All log macros operate similar to the println! macro, accepting a string argument and optionally some variables to be substituted into the message.

Examples

Display is like a log aware print, the data will always appear in the console without any timestamp or other decoration and will be written to the log file with timestamp info.

display!("Some msg");       // This one is equivalent to print! but appears in the log
displayln!("Some msg");     // Like println!

let x = 10;
displayln!("The value of x is {}", x);

// There are also red/green versions of display:
display_green!("The value of x is {}", x);
display_redln!("The value of x is {}", x);

Similar macros exist to display messages at different log levels, all of these appear with timestamps in both the console and the log file:

let x = 10;

log_error!("The value of x is {}", x);
log_info!("The value of x is {}", x);
log_success!("The value of x is {}", x);
log_warning!("The value of x is {}", x);
log_deprecated!("The value of x is {}", x);
log_debug!("The value of x is {}", x);
log_trace!("The value of x is {}", x);

Log Levels

The log verbosity mapping is shown below, level 0 is equivalent to running with no verbosity switch, level 1 is -v, level 2 is -vv, etc.

Level 0 - display, log_error

Level 1 - log_info, log_success, log_warning, log_deprecated

Level 2 - log_debug

Level 3 - log_trace

Structs

Inner

Logger

Functions

log_dir