Module origen::core::config

Collects the main Origen configuration options from all origen.toml files found in the application and Origen CLI installation file system paths

Examples

use origen::ORIGEN_CONFIG;

println!("Server: {}", &ORIGEN_CONFIG.pkg_server);  // => "Server: https://pkgs.company.net:9292"

Structs

CONFIG

Config

Statics

DEFAULT_ENCRYPTION_KEY
Default keys generated from crate::utility::mod::tests::check_default_key_values default_encryption_key: !<<<---Origen StandardKey--->>>!

DEFAULT_ENCRYPTION_NONCE
default_encryption_nonce: ORIGEN NONCE