Struct origen::core::model::registers::bit::Bit

pub struct Bit {
    pub id: usize,
    pub register_id: usize,
    pub overlay: RwLock<Option<String>>,
    pub overlay_snapshots: RwLock<HashMap<String, Option<String>>>,
    pub state: RwLock<u8>,
    pub device_state: RwLock<u8>,
    pub state_snapshots: RwLock<HashMap<String, u8>>,
    pub access: AccessType,
    pub position: usize,
}

Fields
id: usizeregister_id: usizeoverlay: RwLock<Option<String>>overlay_snapshots: RwLock<HashMap<String, Option<String>>>state: RwLock<u8>

The individual bits mean the following: 0 - Data value 1 - Value is X 2 - Value is Z 3 - Bit is to be verified 4 - Bit is to be captured 5 - Modified, sets if bits [2:0] have been changed since the last reset

device_state: RwLock<u8>

The state we think the device has, only bits [2:0] are applicable. This is updated by reseting the register or executing a transaction.

state_snapshots: RwLock<HashMap<String, u8>>

The state of the bit at the last reset

access: AccessTypeposition: usize
Methods

impl Bit

pub fn snapshot(&self, name: &str)

pub fn is_changed(&self, name: &str) -> Result<bool>

pub fn rollback(&self, name: &str) -> Result<()>

pub fn copy_state(&self, source: &Bit)

Copies the state (data and flags) and overlay attributes of the given bit to self

pub fn clear_flags(&self)

pub fn clear_verify_flag(&self)

pub fn capture(&self)

pub fn set_undefined(&self)

Sets the bit's data value to X

pub fn set_hiz(&self)

Sets the bit's data value to Z

pub fn has_known_value(&self) -> bool

Returns true if not in X or Z state

pub fn is_x(&self) -> bool

pub fn is_z(&self) -> bool

pub fn is_to_be_verified(&self) -> bool

pub fn is_to_be_captured(&self) -> bool

pub fn has_overlay(&self) -> bool

pub fn is_readable(&self) -> bool

pub fn is_writeable(&self) -> bool

pub fn is_writable(&self) -> bool

pub fn is_modified_since_reset(&self) -> bool

pub fn state_char(&self) -> char

pub fn verify(&self) -> Result<()>

pub fn is_update_required(&self) -> bool

Returns true if the current bit state differs from the device state. Note that for the purposes of this comparison, X and a 1/0 are considered different.

pub fn update_device_state(&self) -> Result<()>

pub fn data(&self) -> Result<u8>

pub fn set_data(&self, val: u8)

pub fn force_data(&self, val: u8)

Like set_data(), but will force the data value in the event of the bit being unimplemented or otherwise unwritable

pub fn get_overlay(&self) -> Option<String>

pub fn set_overlay(&self, val: Option<&str>)

pub fn verify_enable_flag(&self) -> u8

pub fn capture_enable_flag(&self) -> u8

pub fn overlay_enable_flag(&self) -> u8

pub fn reset(&self, val: u8)

Applies the given state value, making it the new reset baseline and the device_state