Struct origen::core::model::Model

pub struct Model {
    pub id: usize,
    pub name: String,
    pub parent_id: Option<usize>,
    pub sub_blocks: IndexMap<String, usize>,
    pub memory_maps: IndexMap<String, usize>,
    pub pins: IndexMap<String, usize>,
    pub pin_groups: IndexMap<String, usize>,
    pub pin_headers: IndexMap<String, usize>,
    pub timesets: IndexMap<String, usize>,
    pub address_unit_bits: u32,
    pub offset: u128,
    pub services: IndexMap<String, usize>,
}

Fields
id: usizename: Stringparent_id: Option<usize>

The only one without a parent is the top-level DUT model

sub_blocks: IndexMap<String, usize>

All children of this block/model, which are themselves models

memory_maps: IndexMap<String, usize>

All registers owned by this model are arranged within memory maps

pins: IndexMap<String, usize>pin_groups: IndexMap<String, usize>pin_headers: IndexMap<String, usize>timesets: IndexMap<String, usize>address_unit_bits: u32

Represents the number of bits in an address increment between two consecutive addressable units within the block. Its value defaults to 8 indicating that the address offsets of its sub-blocks will be expressed as byte addresses. This attribute can be overridden by individual MemoryMaps defined within the block. Since this attribute is intrinsically linked to the address definitions made within the block, it cannot be overridden by instantiation.

offset: u128

The starting address of the block expressed in address_unit_bits of the parent block. This defaults to 0 but can be overridden by instantiation.

services: IndexMap<String, usize>
Implementations

impl Model

pub fn register_pin_header(
&mut self,
pid: usize,
name: &str,
pin_names: Vec<String>
) -> Result<PinHeader, Error>

pub fn get_pin_header_id(&self, name: &str) -> Option<usize>

impl Model

pub fn register_pin(
&mut self,
pin_group_id: usize,
physical_pin_id: usize,
name: &str,
reset_data: Option<u32>,
reset_action: Option<PinActions>,
endianness: Option<Endianness>
) -> Result<(PinGroup, Pin), Error>

pub fn register_pin_group(
&mut self,
pin_group_id: usize,
name: &str,
pins: Vec<String>,
endianness: Option<Endianness>
) -> Result<PinGroup, Error>

pub fn get_pin_id(&self, name: &str) -> Option<usize>

pub fn get_pin_group_id(&self, name: &str) -> Option<usize>

impl Model

pub fn add_timeset(
&mut self,
model_id: usize,
instance_id: usize,
name: &str,
period: Option<Box<dyn ToString>>,
default_period: Option<f64>
) -> Result<Timeset, Error>

pub fn get_timeset_id(&self, name: &str) -> Option<usize>

pub fn contains_timeset(&self, name: &str) -> bool

impl Model

pub fn new(
id: usize,
name: String,
parent_id: Option<usize>,
offset: Option<u128>
) -> Model

pub fn add_service(&mut self, name: &str, id: usize) -> Result<()>

pub fn lookup(&self, key: &str) -> Result<&IndexMap<String, usize>>

pub fn console_header(&self, dut: &MutexGuard<Dut>) -> (String, usize)

Returns the hierarchical name of the model and the offset for console displays

pub fn get_memory_map_id(&self, name: &str) -> Result<usize>

Get the ID for the given memory map name, throw an error if it doesn't exist

pub fn display_path(&self, dut: &MutexGuard<Dut>) -> String

Returns the path to this model for displaying to a user, e.g. in error messages.

pub fn console_display(&self, dut: &MutexGuard<Dut>) -> Result<String>

pub fn parent<'a>(&self, dut: &'a MutexGuard<Dut>) -> Result<Option<&'a Model>>

Returns the parent of this model or None, normally meaning that this is the top-level model

pub fn address(&self, dut: &MutexGuard<Dut>) -> Result<u128>

Returns the fully resolved address of the block which is comprised of the sum of it's own offset and that of it's parent(s).

pub fn bit_address(&self, dut: &MutexGuard<Dut>) -> Result<u128>

Returns the fully-resolved address taking into account all base addresses defined by the parent hierachy. The returned address is with an address_unit_bits size of 1.

pub fn friendly_path(&self, dut: &MutexGuard<Dut>) -> Result<String>

Returns a path to this block like "dut.my_block"