Trait origen::core::tester::Interceptor

pub trait Interceptor {
    fn cycle(
        &mut self, 
        _repeat: u32, 
        _compressable: bool, 
        _node: &Node
    ) -> Result<()> { ... }

    fn set_timeset(&mut self, _timeset_id: usize, _node: &Node) -> Result<()> { ... }

    fn clear_timeset(&mut self, _node: &Node) -> Result<()> { ... }

    fn cc(&mut self, _level: u8, _msg: &str, _node: &Node) -> Result<()> { ... }

    fn set_pin_header(
        &mut self, 
        _pin_header_id: usize, 
        _node: &Node
    ) -> Result<()> { ... }

    fn clear_pin_header(&mut self, _node: &Node) -> Result<()> { ... }
}

Trait which allows Rust-side implemented testers to intercept generic calls from the tester. Each method will be given the resulting node after processing. Note: the node given is only a clone of what will be stored in the AST.

Provided methods

fn cycle(
&mut self,
_repeat: u32,
_compressable: bool,
_node: &Node
) -> Result<()>

fn set_timeset(&mut self, _timeset_id: usize, _node: &Node) -> Result<()>

fn clear_timeset(&mut self, _node: &Node) -> Result<()>

fn cc(&mut self, _level: u8, _msg: &str, _node: &Node) -> Result<()>

fn set_pin_header(&mut self, _pin_header_id: usize, _node: &Node) -> Result<()>

fn clear_pin_header(&mut self, _node: &Node) -> Result<()>

Loading content...