Trait origen::core::tester::TesterAPI

pub trait TesterAPI: Debug + Interceptor + TesterID + TesterAPIClone {
    fn render_pattern(&mut self, ast: &Node) -> Result<Vec<PathBuf>> { ... }

    fn render_program(&mut self) -> Result<Vec<PathBuf>> { ... }

    fn pattern_differ(
        &self, 
        pat_a: &Path, 
        pat_b: &Path
    ) -> Option<Box<dyn Differ>> { ... }

    fn program_differ(
        &self, 
        pat_a: &Path, 
        pat_b: &Path
    ) -> Option<Box<dyn Differ>> { ... }

    fn pin_action_resolver(&self) -> Option<Resolver> { ... }

    fn output_dir(&self) -> Result<PathBuf> { ... }
}

Provided methods

fn render_pattern(&mut self, ast: &Node) -> Result<Vec<PathBuf>>

Render the given AST to an output, returning the path(s) to the created file(s) if successful. A default implementation is given since some testers may only support prog gen and not patgen and vice versa, in that case they will return an empty vector.

fn render_program(&mut self) -> Result<Vec<PathBuf>>

Render the test program to an output, returning the path(s) to the created file(s) if successful. A default implementation is given since some testers may only support prog gen and not patgen and vice versa, in that case they will return an empty vector.

fn pattern_differ(&self, pat_a: &Path, pat_b: &Path) -> Option<Box<dyn Differ>>

The tester should implement this to return a differ instance which is configured per the tester's pattern format, e.g. to define the command char(s). If diff'ing is not applicable to the tester, e.g. the pattern is in binary format, then the tester does not need to implement this. If only some patterns can be diffed then then test should return None in the case where the pattern is one that cannot be diffed.

fn program_differ(&self, pat_a: &Path, pat_b: &Path) -> Option<Box<dyn Differ>>

fn pin_action_resolver(&self) -> Option<Resolver>

fn output_dir(&self) -> Result<PathBuf>

Returns a path to the tester-specific output directory, it is expected to create it if it doesn't exist so the caller doesn't have to

Loading content...