Trait origen::core::tester::TesterAPI

pub trait TesterAPI: Debug + Interceptor {
    fn name(&self) -> String;

    fn clone(&self) -> Box<dyn TesterAPI + Send>;

    fn render_pattern(&mut self, ast: &Node) -> Result<Vec<PathBuf>> { ... }

    fn to_string(&self) -> String { ... }

    fn pattern_differ(&self, pat_a: &Path, pat_b: &Path) -> Option<Differ> { ... }
}

Required methods

fn name(&self) -> String

fn clone(&self) -> Box<dyn TesterAPI + Send>

Loading content...
Provided methods

fn render_pattern(&mut self, ast: &Node) -> Result<Vec<PathBuf>>

Render the given AST to an output, returning the path(s) to the created file(s) if successfull. A default implementation is given since some testers may only support prog gen and not patgen and vice versa, in that case they will return an empty vector.

fn to_string(&self) -> String

fn pattern_differ(&self, pat_a: &Path, pat_b: &Path) -> Option<Differ>

The tester should implement this to return a differ instance which is configured per the tester's pattern format, e.g. to define the command char(s). If diff'ing is not applicable to the tester, e.g. the pattern is in binary format, then the tester does not need to implement this. If only some patterns can be diffed then then test should return None in the case where the pattern is one that cannot be diffed.

Loading content...