Module origen::utility::file_utils

Functions

_copy

copy
Copy the given file or directory to the given directory, directories will be copied recursively.

copy_contents
Like copy, however if the source is a directory then its contents will be copied to the target directory, rather than copying the source folder itself

mv
Move a file or directory

symlink
Create a symlink, works on Linux or Windows. The dst path will be a symbolic link pointing to the src path.

to_relative_path
Returns the given abs path as a relative path. By default it will be made relative to the PWD, but an alternative path to make it relative to can be supplied. Can return an error if

with_dir
Temporarily sets the current dir to the given dir for the duration of the given function and then restores it at the end. An error will be returned if there is a problem switching to the given directory, e.g. if it doesn't exist, otherwise the result from the given function is returned.